modified entitlement wizard logic
