+++ g2p_program_cycleless/__manifest__.py
@@ -12,10 +12,11 @@
         "g2p_programs",
     ],
     "data": [
+        "security/ir.model.access.csv",
         "views/program_view.xml",
         "views/program_membership.xml",
         "views/program_manager_view.xml",
-        "views/entitlement.xml",
+        "wizard/create_entitlement_wizard.xml",
     ],
     "assets": {},
     "demo": [],
