+++ g2p_program_cycleless/models/__init__.py
@@ -1,6 +1,5 @@
-# Part of OpenSPP. See LICENSE file for full copyright and licensing details.
+# Part of OpenG2P. See LICENSE file for full copyright and licensing details.
 from . import program_manager
 from . import program_membership
 from . import entitlement_manager
 from . import programs
-from . import entitlement
