+++ g2p_program_cycleless/wizard/__init__.py
@@ -1,3 +1,3 @@
 # Part of OpenG2P. See LICENSE file for full copyright and licensing details.
-from . import models
-from . import wizard
+
+from . import create_entitlement_wizard
